헤더 라인 파싱 과정에서 들어가는 빈칸 처리 및 오류 수정
